query_construction/self_query/pyproject.toml
[tool.poetry]
name = "self-query"
version = "0.1.0"
description = ""
authors = ["Anantvir Singh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.15"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.15"
python-dotenv = "^1.0.1"
lark = "^1.2.2"
langchain-chroma = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
